Add ingredient to bar, custom cocktails, popular cocktails

+++ src/main/java/ru/vsu/amm/inshaker/services/CocktailService.java
@@ -2,6 +2,9 @@ package ru.vsu.amm.inshaker.services;
 
 import org.dozer.Mapper;
 import org.springframework.beans.BeanUtils;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import ru.vsu.amm.inshaker.exceptions.AnonymousAuthenticationException;
 import ru.vsu.amm.inshaker.exceptions.CocktailNotFoundException;
@@ -41,28 +44,25 @@ public class CocktailService {
         this.cocktailDTOConverter = cocktailDTOConverter;
     }
 
-    public List<Cocktail> getAllCocktails(String search, String base, String spirit, String group, List<String> tastes) {
-        if (search == null && base == null && spirit == null && group == null && tastes == null) {
-            return cocktailRepository.findAll();
-        }
-
-        Spirit s = Spirit.findByRuName(spirit);
-        if (tastes == null) {
-            return cocktailRepository.findAllWithFilters(search, base, s.getRangeLow(), s.getRangeHigh(), group);
-        } else {
-            return cocktailRepository.findAllWithFilters(search, base, s.getRangeLow(), s.getRangeHigh(), group, tastes, (long) tastes.size());
-        }
-    }
-
-    public List<CocktailSimpleDTO> getAll(String search, String base, String spirit, String group, List<String> tastes) {
-        return getAllCocktails(search, base, spirit, group, tastes).stream()
-                .map(c -> mapper.map(c, CocktailSimpleDTO.class)).collect(Collectors.toList());
-    }
 
     public Cocktail getCocktail(Long id) {
         return cocktailRepository.findById(id)
                 .orElseThrow(() -> new CocktailNotFoundException(id));
     }
+
+    public List<Cocktail> getAllCocktails(User author, String search, String base, String spirit, String group, List<String> tastes) {
+        if (author == null && search == null && base == null && spirit == null && group == null && tastes == null) {
+            return cocktailRepository.findAllByAuthorIsNull();
+        }
+
+        Spirit s = Spirit.findByRuName(spirit);
+        if (tastes == null) {
+            return cocktailRepository.findAllWithFilters(author, search, base, s.getRangeLow(), s.getRangeHigh(), group);
+        } else {
+            return cocktailRepository.findAllWithFilters(author, search, base, s.getRangeLow(), s.getRangeHigh(), group, tastes, (long) tastes.size());
+        }
+    }
+
 
     public CocktailDTO get(Long id) {
         Cocktail cocktail = getCocktail(id);
@@ -77,6 +77,42 @@ public class CocktailService {
 
         return cocktailDTO;
     }
+
+    public List<CocktailSimpleDTO> getAll(String search, String base, String spirit, String group, List<String> tastes) {
+        return getAllCocktails(null, search, base, spirit, group, tastes).stream()
+                .map(c -> mapper.map(c, CocktailSimpleDTO.class)).collect(Collectors.toList());
+    }
+
+    public List<CocktailSimpleDTO> getPopular(int limit) {
+        return userRepository.findPopularCocktails(PageRequest.of(0, limit)).stream()
+                .map(c -> mapper.map(c, CocktailSimpleDTO.class)).collect(Collectors.toList());
+    }
+
+
+    public CocktailDTO add(CocktailDTO cocktail) {
+        Cocktail newCocktail = cocktailDTOConverter.convert(cocktail);
+        newCocktail.setId(null);
+        newCocktail.setAuthor(null);
+        return mapper.map(cocktailRepository.save(newCocktail), CocktailDTO.class);
+    }
+
+    public CocktailDTO update(CocktailDTO newCocktail, Long id) {
+        return cocktailRepository.findById(id)
+                .map(oldCocktail -> {
+                    BeanUtils.copyProperties(cocktailDTOConverter.convert(newCocktail), oldCocktail);
+                    oldCocktail.setId(id);
+                    return mapper.map(cocktailRepository.save(oldCocktail), CocktailDTO.class);
+                }).orElseThrow(() -> new CocktailNotFoundException(id));
+    }
+
+    public void delete(Long id) {
+        try {
+            cocktailRepository.deleteById(id);
+        } catch (EmptyResultDataAccessException ex) {
+            throw new CocktailNotFoundException(id);
+        }
+    }
+
 
     public Set<String> getBases() {
         return cocktailRepository.findDistinctBases();
@@ -93,6 +129,7 @@ public class CocktailService {
     public Set<String> getSpirits() {
         return Arrays.stream(Spirit.values()).map(Spirit::getRuName).collect(Collectors.toSet());
     }
+
 
     public Set<CocktailSimpleDTO> getFavorites() {
         return userService.getCurrentUser().getFavorite().stream()
@@ -111,23 +148,43 @@ public class CocktailService {
         userRepository.save(currentUser);
     }
 
-    public CocktailDTO add(CocktailDTO cocktail) {
+
+    public CocktailDTO getCustom(Long id) {
+        Cocktail cocktail = getCocktail(id);
+        if (cocktail.getAuthor() == userService.getCurrentUser()) {
+            return mapper.map(cocktail, CocktailDTO.class);
+        } else throw new AccessDeniedException("The user does not have permission to get the cocktail");
+    }
+
+    public List<CocktailSimpleDTO> getAllCustoms(String search, String base, String spirit, String group, List<String> tastes) {
+        return getAllCocktails(userService.getCurrentUser(), search, base, spirit, group, tastes).stream()
+                .map(c -> mapper.map(c, CocktailSimpleDTO.class)).collect(Collectors.toList());
+    }
+
+
+    public CocktailDTO addCustom(CocktailDTO cocktail) {
         Cocktail newCocktail = cocktailDTOConverter.convert(cocktail);
         newCocktail.setId(null);
+        newCocktail.setAuthor(userService.getCurrentUser());
         return mapper.map(cocktailRepository.save(newCocktail), CocktailDTO.class);
     }
 
-    public CocktailDTO update(CocktailDTO newCocktail, Long id) {
-        return cocktailRepository.findById(id)
-                .map(oldCocktail -> {
-                    BeanUtils.copyProperties(cocktailDTOConverter.convert(newCocktail), oldCocktail);
-                    oldCocktail.setId(id);
-                    return mapper.map(cocktailRepository.save(oldCocktail), CocktailDTO.class);
-                }).orElseThrow(() -> new CocktailNotFoundException(id));
+    public CocktailDTO updateCustom(CocktailDTO newCocktail, Long id) {
+        Cocktail oldCocktail = getCocktail(id);
+        User currentUser = userService.getCurrentUser();
+        if (oldCocktail.getAuthor() == currentUser) {
+            BeanUtils.copyProperties(cocktailDTOConverter.convert(newCocktail), oldCocktail);
+            oldCocktail.setId(id);
+            oldCocktail.setAuthor(currentUser);
+            return mapper.map(cocktailRepository.save(oldCocktail), CocktailDTO.class);
+        } else throw new AccessDeniedException("The user does not have permission to update the cocktail");
     }
 
-    public void delete(Long id) {
-        cocktailRepository.deleteById(id);
+    public void deleteCustom(Long id) {
+        Cocktail cocktail = getCocktail(id);
+        if (cocktail.getAuthor() == userService.getCurrentUser()) {
+            cocktailRepository.deleteById(id);
+        } else throw new AccessDeniedException("The user does not have permission to delete the cocktail");
     }
 
 }
